build: re-organize files so that they also work inside Docker
