+++ .eslintrc.json
@@ -1,7 +1,7 @@
 {
     "parser": "@typescript-eslint/parser",
     "parserOptions": {
-        "project": ["tsconfig.json", "src/tsconfig.json"],
+        "project": ["tsconfig.json", "test/tsconfig.json"],
         "tsconfigRootDir": "."
     },
     "plugins": ["@typescript-eslint", "prettier", "jest"],
