+++ .gitignore
@@ -1,5 +1,10 @@
 coverage
-dist
+bin/*.d.ts
+bin/*.js
+lib/*.d.ts
+lib/*.js
 node_modules
 
 .npmrc
+index.d.ts
+index.js
