+++ Dockerfile
@@ -1,5 +1,7 @@
 FROM node:alpine as builder
 
+WORKDIR /teams-logger
+
 COPY . .
 RUN npm ci
 RUN npm run build
@@ -7,12 +9,11 @@ RUN npm run build
 FROM node:alpine
 
 LABEL org.opencontainers.image.source=https://github.com/verkkokauppacom/teams-logger
-ENTRYPOINT [ "/teams-logger/bin.js" ]
-
-RUN mkdir /teams-logger
 WORKDIR /teams-logger
+ENTRYPOINT ["node" ,"bin/teams-logger.js"]
 
-COPY --from=builder /package.json /package-lock.json .
+COPY --from=builder /teams-logger/package.json /teams-logger/package-lock.json /teams-logger/
 RUN npm ci --production && rm package-lock.json
 
-COPY --from=builder /dist/* .
+COPY --from=builder /teams-logger/bin/*.js /teams-logger/bin/
+COPY --from=builder /teams-logger/lib/*.js /teams-logger/lib/
