+++ index.ts
@@ -1,0 +1,2 @@
+export { default as rawLogger } from './lib/rawLogger'
+export { default as simpleLogger } from './lib/simpleLogger'
