+++ lib/addPotentialActions.ts
