+++ lib/coerceJson.ts
