+++ lib/coerceLinks.ts
