+++ lib/commandDefault.ts
