+++ lib/commandRaw.ts
