+++ lib/rawLogger.ts
