+++ lib/simpleLogger.ts
