+++ package.json
@@ -11,18 +11,19 @@
         "node": ">= 10"
     },
     "bin": {
-        "teams-logger": "bin.js"
+        "teams-logger": "bin/teams-logger.js"
     },
     "files": [
-        "*.d.ts",
-        "*.js"
+        "lib/**/*.d.ts",
+        "lib/**/*.js",
+        "index.d.ts",
+        "index.js"
     ],
     "types": "index.d.ts",
     "scripts": {
         "lint:base": "eslint --ext .js,.ts",
         "lint": "npm run lint:base .",
-        "build": "tsc --build src",
-        "postbuild": "npm-publish-files --force dist",
+        "build": "tsc --build",
         "test": "jest --coverage",
         "version": "standard-version --no-verify"
     },
@@ -42,7 +43,7 @@
         }
     },
     "lint-staged": {
-        "*.{js,json}": "lint:base --fix"
+        "*.{js,ts}": "npm run lint:base -- --fix"
     },
     "standard-version": {
         "scripts": {
@@ -74,7 +75,6 @@
         "husky": "^4.3.8",
         "jest": "26.6.3",
         "lint-staged": "^10.5.4",
-        "npm-publish-files": "1.3.0",
         "prettier": "^2.2.1",
         "standard-version": "^9.1.1",
         "ts-jest": "26.5.1",
