--- src/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-        "outDir": "../dist",
-        "noEmit": false
-    },
-    "exclude": [
-        "../test"
-    ]
-}
