+++ test/coerceJson.spec.ts
@@ -1,4 +1,5 @@
-import coerceJson from '../src/coerceJson'
+/** @ts-expect-error - explicit .ts import for test coverage */
+import coerceJson from '../lib/coerceJson.ts'
 
 describe('coerceJson', () => {
     it('should parse valid JSON', () => {
@@ -8,7 +9,6 @@ describe('coerceJson', () => {
     })
 
     it('should throw when invalid JSON', () => {
-        /** @ts-expect-error - invalid argument */
         expect(() => coerceJson()).toThrowErrorMatchingInlineSnapshot(
             `"Unable to parse JSON!"`
         )
