+++ tsconfig.json
@@ -4,7 +4,6 @@
     "esModuleInterop": true,
     "lib": ["es6"],
     "moduleResolution": "node",
-    "noEmit": true,
     "noResolve": false,
     "noUnusedParameters": true,
     "pretty": true,
@@ -13,7 +12,9 @@
     "target": "es5"
   },
   "include": [
-    "src/**/*.ts",
-    "test/**/*.ts"
+    "bin/**/*.ts",
+    "lib/**/*.ts",
+    "types/**/*.d.ts",
+    "index.ts"
   ]
 }
